+++ opensora/models/ae/videobase/__init__.py
@@ -5,14 +5,13 @@ from .vqvae import (
     VQVAEConfiguration,
     VQVAEModel,
     VQVAETrainer,
-    VQVAEDataset,
+    VQVAEDataset, VideoGPTVQVAEWrapper
 )
 
 
 """
 old code
 """
-from .vqvae.videogpt import VideoGPTVQVAEWrapper
 
 videovqvae = [
     "bair_stride4x2x2",
@@ -23,6 +22,7 @@ videovqvae = [
 videovae = []
 
 videobase_ae_stride = {
+    'checkpoint-14000': [4, 4, 4],
     'bair_stride4x2x2': [4, 2, 2],
     'ucf101_stride4x4x4': [4, 4, 4],
     'kinetics_stride4x4x4': [4, 4, 4],
@@ -30,6 +30,7 @@ videobase_ae_stride = {
 }
 
 videobase_ae_channel = {
+    'checkpoint-14000': 4,
     'bair_stride4x2x2': 256,
     'ucf101_stride4x4x4': 256,
     'kinetics_stride4x4x4': 256,
@@ -37,6 +38,7 @@ videobase_ae_channel = {
 }
 
 videobase_ae = {
+    "checkpoint-14000": VideoGPTVQVAEWrapper,
     "bair_stride4x2x2": VideoGPTVQVAEWrapper,
     "ucf101_stride4x4x4": VideoGPTVQVAEWrapper,
     "kinetics_stride4x4x4": VideoGPTVQVAEWrapper,
