+++ opensora/train/train.py
@@ -56,9 +56,10 @@ def main(args):
         experiment_dir = f"{args.results_dir}/{args.resume_from_checkpoint}"
     else:
         experiment_index = len(glob(f"{args.results_dir}/*"))
-        model_string_name = args.model.replace("/", "-")  # e.g., Latte-XL/2 --> Latte-XL-2 (for naming folders)
+        model_string_name = args.model.replace("/", "-")  # e.g., Latte-XL/122 --> Latte-XL-122 (for naming folders)
+        ae_string_name = args.ae.split("/")[-1]  # e.g., stabilityai/sd-vae-ft-mse --> sd-vae-ft-mse (for naming folders)
         num_frame_string = 'F' + str(args.num_frames) + 'S' + str(args.sample_rate)
-        experiment_dir = f"{args.results_dir}/{experiment_index:03d}-{model_string_name}-{num_frame_string}-{args.dataset}"  # Create an experiment folder
+        experiment_dir = f"{args.results_dir}/{experiment_index:03d}-{model_string_name}-{ae_string_name}-{num_frame_string}-{args.dataset}"  # Create an experiment folder
         experiment_dir = get_experiment_dir(experiment_dir, args)
     checkpoint_dir = f"{experiment_dir}/checkpoints"  # Stores saved model checkpoints
     if accelerator.is_main_process:
@@ -68,7 +69,6 @@ def main(args):
         tb_writer = create_tensorboard(experiment_dir)
         logger.info(f"Experiment directory created at {experiment_dir}")
         logger.info(f'{args}')
-        logger.info(f'{accelerator}')
     else:
         logger = create_logger(None)
         tb_writer = None
@@ -81,10 +81,10 @@ def main(args):
     patch_size_t, patch_size_h, patch_size_w = int(patch_size[0]), int(patch_size[1]), int(patch_size[2])
     args.patch_size = patch_size_h
     args.patch_size_t, args.patch_size_h, args.patch_size_w = patch_size_t, patch_size_h, patch_size_w
-    assert args.num_frames % ae_stride_t == 0, "Image size must be divisible by 8 (for the VAE encoder)."
-    assert args.max_image_size % ae_stride_h == 0, "Image size must be divisible by 8 (for the VAE encoder)."
-    assert ae_stride_h == ae_stride_w, "Support now."
-    assert patch_size_h == patch_size_w, "Support now."
+    assert ae_stride_h == ae_stride_w, f"Support only ae_stride_h == ae_stride_w now, but found ae_stride_h ({ae_stride_h}), ae_stride_w ({ae_stride_w})"
+    assert patch_size_h == patch_size_w, f"Support only patch_size_h == patch_size_w now, but found patch_size_h ({patch_size_h}), patch_size_w ({patch_size_w})"
+    assert args.num_frames % ae_stride_t == 0, f"Num_frames must be divisible by ae_stride_t, but found num_frames ({args.num_frames}), ae_stride_t ({ae_stride_t})."
+    assert args.max_image_size % ae_stride_h == 0, f"Image size must be divisible by ae_stride_h, but found max_image_size ({args.max_image_size}),  ae_stride_h ({ae_stride_h})."
 
     latent_size = (args.max_image_size // ae_stride_h, args.max_image_size // ae_stride_w)
 
@@ -100,31 +100,21 @@ def main(args):
 
     # # use pretrained model?
     if args.pretrained:
-        checkpoint = torch.load(args.pretrained, map_location='cpu')
-        if "ema" in checkpoint:  # supports checkpoints from train.py
-            logger.info('Using ema ckpt!')
-            checkpoint = checkpoint["ema"]
-        elif "model" in checkpoint:  # supports checkpoints from train.py
-            logger.info('Using model ckpt!')
-            checkpoint = checkpoint["model"]
-        if 'temp_embed' in checkpoint:
-            del checkpoint['temp_embed']
-        if 'pos_embed' in checkpoint:
-            del checkpoint['pos_embed']
-        model_dict = model.state_dict()
-        # 1. filter out unnecessary keys
-        pretrained_dict = {}
-        for k, v in checkpoint.items():
-            if k in model_dict:
-                pretrained_dict[k] = v
-            else:
-                logger.info('Ignoring: {}'.format(k))
-        logger.info('Successfully Load {}% original pretrained model weights '.format(len(pretrained_dict) / len(checkpoint.items()) * 100))
-        # 2. overwrite entries in the existing state dict
-        model_dict.update(pretrained_dict)
-        msg = model.load_state_dict(model_dict, strict=False)
-        logger.info('Successfully load model at {}!'.format(args.pretrained))
-        logger.info(f'{msg}')
+        # load from pixart-alpha
+        pixelart_alpha = torch.load(args.pretrained, map_location='cpu')['state_dict']
+        checkpoint = {}
+        for k, v in pixelart_alpha.items():
+            if 'x_embedder' in k or 't_embedder' in k:
+                checkpoint[k] = v
+            if k.startswith('blocks'):
+                if 'cross_attn' not in k and 'drop_path' not in k and 'scale_shift_table' not in k:
+                    if 'q_norm' not in k and 'k_norm' not in k:
+                        blk_id = k[7]
+                        new_k = f'blocks.{str(int(blk_id) * 2)}{k[8:]}'
+                        checkpoint[new_k] = v
+        total_load = set(model.state_dict().keys()) & set(checkpoint.keys())
+        msg = model.load_state_dict(checkpoint, strict=False)
+        logger.info(f'Successfully load {len(total_load)} keys from {args.pretrained}!')
 
     model = model.to(device)
     # Note that parameter initialization is done within the DiT constructor
@@ -137,11 +127,12 @@ def main(args):
         model = torch.compile(model)
 
     ae = getae(args).to(device)
-    logger.info(f"{model}")
-    logger.info(f"Model Parameters: {sum(p.numel() for p in model.parameters()):,}")
-    for n, p in model.named_parameters():
-        if p.requires_grad:
-            logger.info(f"Training Parameters: {n}")
+    if accelerator.is_main_process:
+        logger.info(f"{model}")
+        logger.info(f"Model Parameters: {sum(p.numel() for p in model.parameters()):,}")
+        for n, p in model.named_parameters():
+            if p.requires_grad:
+                logger.info(f"Training Parameters: {n}")
 
     # Setup optimizer (we used default Adam betas=(0.9, 0.999) and a constant learning rate of 1e-4 in our paper):
     opt = torch.optim.AdamW(model.parameters(), lr=args.lr, weight_decay=0)
@@ -189,7 +180,6 @@ def main(args):
 
     # Potentially load in the weights and states from a previous save
     if args.resume_from_checkpoint:
-        # TODO, need to checkout
         # Get the most recent checkpoint
         dirs = os.listdir(os.path.join(experiment_dir, 'checkpoints'))
         dirs = [d for d in dirs if d.endswith("pt")]
@@ -236,9 +226,9 @@ def main(args):
             if args.extras == 78: # text-to-video
                 raise 'T2V training are Not supported at this moment!'
             elif args.extras == 2:
-                model_kwargs = dict(y=y, attention_mask=attn_mask)
+                model_kwargs = dict(y=y, attn_mask=attn_mask)
             else:
-                model_kwargs = dict(y=None, attention_mask=attn_mask)
+                model_kwargs = dict(y=None, attn_mask=attn_mask)
             with torch.autocast(device_type='cuda', dtype=dtype):
                 t = torch.randint(0, diffusion.num_timesteps, (x.shape[0],), device=device)
                 loss_dict = diffusion.training_losses(model, x, t, model_kwargs)
@@ -321,10 +311,7 @@ if __name__ == "__main__":
 
 
     # --------------------------------------
-    parser.add_argument("--ae", type=str, choices=['bair_stride4x2x2', 'ucf101_stride4x4x4',
-                                                    'kinetics_stride4x4x4', 'kinetics_stride2x4x4',
-                                                   'stabilityai/sd-vae-ft-mse', 'stabilityai/sd-vae-ft-ema'],
-                        default="ucf101_stride4x4x4")
+    parser.add_argument("--ae", type=str, default="ucf101_stride4x4x4")
     parser.add_argument("--extras", type=int, default=2, choices=[1, 2, 78])
     parser.add_argument("--pretrained", type=str, default=None)
     parser.add_argument("--sample-rate", type=int, default=4)
